Allow admin user/password through environment vars.
Allows specifying the default admin user / password through the `PI_ADMIN_USER`
/ `PI_ADMIN_PASSWORD` (or `PI_ADMIN_USER_FILE` / `PI_ADMIN_PASSWORD_FILE`)
environment variables, and update the documentation to strongly recommend
changing the default password.
(The reference to the possibility to remove the default admin user has been
removed, as it allows the possibility of an insecure setup if the database
were to be recreated).

Fixes: #220
Signed-off-by: Joan Bruguera <joanbrugueram@gmail.com>